requirements.txt
python==3.7.15
numpy==1.21.5
matplotlib==3.5.2
pytorch==1.13.1
pytorch-cuda==11.7
scipy==1.7.3
scikit-learn==1.0.2
joblib==1.1.1